requirements.txt
--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

python-dateutil==1.4.1
python-openid==2.2.4
python-yadis==1.1.0
simplejson==2.0.9
pytz==2009l
docutils==0.5
smartypants==1.6.0.3
Markdown==2.0.1
textile==2.1.3
diff-match-patch==20090804
django-atomformat==0.1.1dev
creole==1.0.1


django-notification==0.1.4
django-openid>=0.2.3.dev0
django-email-confirmation==0.1.4
django-mailer==0.1.0
django-announcements==0.1.0
django-pagination==1.0.5.1
django-timezones>0.2.dev0
django-tagging==0.3.1
django-threadedcomments==0.5.2
django-ajax-validation==0.1.3
django-avatar==1.0.3
django-uni-form==0.7.0
django-wikiapp==0.2.0
django-attachments==0.3
django-markup==0.3
django-sorting>=0.1
django-filter==0.5.2
template_utils==0.4p2
django-staticfiles==0.2.0
django-groups>=0.1-dev1
GitPython

-e git+git://github.com/mintchaos/django_compressor.git#django-compressor
-e hg+http://code.welldev.org/django-oauth#egg=django-oauth
-e git+git://github.com/alex/django-templatetag-sugar.git#egg=django-templatetag-sugar
